+++ classification/utils.py
@@ -261,7 +261,7 @@ def test_single_label(model, logger, test_loader, criterion, classes_names, serv
     with torch.no_grad():
         for batch_idx, (image, batch_label) in enumerate(test_loader):
             image = image.to(device=DEVICE, dtype=torch.float32)
-            batch_label = batch_label.to(device=DEVICE, dtype=torch.float32)
+            batch_label = batch_label.to(device=DEVICE)
 
             logits = model(image)
             loss = criterion(logits, batch_label)
